Fix crash when performing update while animating
The `completion` block is called when the animation has completed, but at the end of the `updates` block the data has already been updated (from the perspective of the `UICollectionView`).

Not calling `reset()` and setting `defersUpdate = false` could cause crashes e.g. due to changes being applied multiple times.

The best way to recreate this is to have a series of updates occur very close to each other but with the "Slow Animation" setting enabled in the simulator.

This change might also apply to table views, I have not checked.